+++ GUI-dev/src/app/pages/pages-routing.module.ts
@@ -3,9 +3,11 @@ import { NgModule } from '@angular/core';
 import { PagesComponent } from './pages.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { HomeComponent } from './home/home.component';
+import {AuthGuard} from '../auth-guard.service';
+import {ApprovalComponent} from './approval/approval.component';
 import {ActiveProcessesComponent} from './dashboard/components/activeProcesses/activeProcesses.component';
 // import {EventLoggerComponent} from './admin/components/eventLogger/eventLogger.component';
-import {AuthGuard} from '../auth-guard.service';
+import { ProcessStoreSearchComponent } from './processstore-search/processstore-search.component';
 import {UserDetailsComponent} from './user-details/user-details.component';
 
 
@@ -25,8 +27,18 @@ const routes: Routes = [{
     },
 
     {
+      path: 'processstore-search',
+      component: ProcessStoreSearchComponent,
+      canActivate: [AuthGuard],
+    },
+    {
       path: 'home',
       component: HomeComponent,
+    },
+    {
+      path: 'approval',
+      component: ApprovalComponent,
+      canActivate: [AuthGuard],
     },
     {
       path: '',