+++ GUI-dev/src/app/pages/pages.module.ts
@@ -4,8 +4,11 @@ import { DashboardModule } from './dashboard/dashboard.module';
 import { PagesRoutingModule } from './pages-routing.module';
 import { ThemeModule } from '../@theme/theme.module';
 import {HomeModule} from './home/home.module';
-import {UserDetailsModule} from './user-details/user-details.module';
+import { ProcessStoreSearchModule } from './processstore-search/processstore-search.module';
+import {ApprovalModule} from './approval/approval.module';
+import {ApprovalDetailsModule} from './approval-details/approval-details.module';
 import AdminModule from './admin/admin.module';
+import {UserDetailsModule} from './user-details/user-details.module';
 
 const PAGES_COMPONENTS = [
   PagesComponent,
@@ -17,8 +20,11 @@ const PAGES_COMPONENTS = [
     ThemeModule,
     DashboardModule,
     HomeModule,
+    ApprovalModule,
+    ApprovalDetailsModule,
     AdminModule,
     UserDetailsModule,
+    ProcessStoreSearchModule,
   ],
   declarations: [
     ...PAGES_COMPONENTS,